+++ tests/001_functionals/suite_009_pagebase.py
@@ -5,7 +5,6 @@
 import pytest
 from qacode.core.testing.test_info import TestInfoBotUnique
 from qacode.core.webs.controls.control_base import ControlBase
-from qacode.core.webs.controls.control_form import ControlForm
 from qacode.core.webs.pages.page_base import PageBase
 from qacode.utils import settings
 
@@ -119,10 +118,5 @@ class TestPageBase(TestInfoBotUnique):
             cfg.get('url'))
         for config_control in self.page.get('controls'):
             name = config_control.get('name')
-            instance_name = config_control.get('instance')
             ctl = page.get_element(config_control)
-            if instance_name == 'ControlBase':
-                self.assert_is_instance(ctl, ControlBase)
-            elif instance_name == 'ControlForm':
-                self.assert_is_instance(ctl, ControlForm)
             self.assert_in(name, dir(page))